+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE = "t4c2-1.8.0";
+const CACHE = "t4c2-1.9.0";
 const ASSETS = [
   "./",
   "./index.html",